+++ CHANGELOG.md
@@ -1,4 +1,3 @@
 # ChangeLog
 
-* fix widget group server crash
-* fix jei compat with only JEI APIs
+* support custom persisted name for IRef